some post merge refactorings
