+++ core/src/main/java/com/mankomania/game/core/network/messages/servertoclient/stock/EndStockMessage.java
@@ -1,4 +1,4 @@
-package com.mankomania.game.core.network.messages.servertoclient.minigames;
+package com.mankomania.game.core.network.messages.servertoclient.stock;
 
 import java.util.Map;
