+++ gamecore/src/main/java/com/mankomania/game/gamecore/util/Screen.java
@@ -1,8 +1,6 @@
 package com.mankomania.game.gamecore.util;
 
-import com.mankomania.game.gamecore.client.NetworkClient;
-
-import com.mankomania.game.gamecore.StockExchange.AktienBoerse;
+import com.mankomania.game.gamecore.screens.StockExchange.AktienBoerse;
 import com.mankomania.game.gamecore.screens.*;
 import com.mankomania.game.gamecore.screens.trickyone.TrickyOneScreen;
 
@@ -35,11 +33,11 @@ public enum Screen {
         public AbstractScreen getScreen(Object... params) {
             return new MainGameScreen();
         }
-         },
-    AKTIEN_BOERSE{
-            public AbstractScreen getScreen(Object... params) {
-                return new AktienBoerse();
-            }
+    },
+    AKTIEN_BOERSE {
+        public AbstractScreen getScreen(Object... params) {
+            return new AktienBoerse();
+        }
     },
     TRICKY_ONE {
         public AbstractScreen getScreen(Object... params) {
